# GH-93657: More LOAD_ATTR specializations