--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -285,8 +285,10 @@ _specializations = {
         "LOAD_ATTR_ADAPTIVE",
         # These potentially push [NULL, bound method] onto the stack.
         "LOAD_ATTR_CLASS",
+        "LOAD_ATTR_CLASS_MUTABLE_DESCRIPTOR",
         "LOAD_ATTR_INSTANCE_VALUE",
         "LOAD_ATTR_MODULE",
+        "LOAD_ATTR_PROPERTY",
         "LOAD_ATTR_SLOT",
         "LOAD_ATTR_WITH_HINT",
         # These will always push [unbound method, self] onto the stack.