--- a/Misc/NEWS.d/next/Core and Builtins/2022-06-16-08-56-26.gh-issue-93657.AZG2WT.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-06-16-08-56-26.gh-issue-93657.AZG2WT.rst
@@ -1,1 +1,1 @@
-Specialize `LOAD_ATTR` for mutable class attribute descriptors and ``property()`` attributes.
+Specialize ``LOAD_ATTR`` for mutable class attribute descriptors and ``property()`` attributes.