--- a/Python/ceval.c
+++ b/Python/ceval.c
@@ -3685,25 +3685,19 @@ handle_eval_breaker:
             uint32_t type_version = read_u32(cache->type_version);
             DEOPT_IF(cls->tp_version_tag != type_version, LOAD_ATTR);
             assert(type_version != 0);
-            PyObject *prop = read_obj(cache->descr);
-            assert(Py_TYPE(prop) == &PyProperty_Type);
-            PyObject *fget = _PyProperty_PropGet(prop);
-            DEOPT_IF(!PyFunction_Check(fget), LOAD_ATTR);
+            PyObject *fget = read_obj(cache->descr);
             PyFunctionObject *f = (PyFunctionObject *)fget;
             DEOPT_IF(f->func_version != cache->keys_version[0], LOAD_ATTR);
             PyCodeObject *code = (PyCodeObject *)f->func_code;
-            size_t size = code->co_nlocalsplus + code->co_stacksize + FRAME_SPECIALS_SIZE;
             assert(code->co_argcount == 1);
             STAT_INC(LOAD_ATTR, hit);
 
-            _PyInterpreterFrame *new_frame = _PyThreadState_BumpFramePointer(tstate, size);
+            Py_INCREF(fget);
+            _PyInterpreterFrame *new_frame = _PyFrame_Push(tstate, f);
             if (new_frame == NULL) {
                 goto error;
             }
-            CALL_STAT_INC(frames_pushed);
-            Py_INCREF(f);
-            _PyFrame_InitializeSpecials(new_frame, f,
-                NULL, code->co_nlocalsplus);
+            CALL_STAT_INC(inlined_py_calls);
             SET_TOP(NULL);
             STACK_SHRINK(!(oparg & 1));
             new_frame->localsplus[0] = owner;