--- a/Python/ceval.c
+++ b/Python/ceval.c
@@ -1465,8 +1465,8 @@ eval_frame_handle_pending(PyThreadState *tstate)
                             assert(STACK_LEVEL() <= STACK_SIZE()); \
                         } while (0)
 #define STACK_SHRINK(n) do { \
-                            assert((n) >= 0); \
-                            assert(STACK_LEVEL() >= (n)); \
+                            assert(n >= 0); \
+                            assert(STACK_LEVEL() >= n); \
                             BASIC_STACKADJ(-(n)); \
                         } while (0)
 #else
@@ -3699,7 +3699,8 @@ handle_eval_breaker:
             }
             CALL_STAT_INC(inlined_py_calls);
             SET_TOP(NULL);
-            STACK_SHRINK(!(oparg & 1));
+            int push_null = !(oparg & 1);
+            STACK_SHRINK(push_null);
             new_frame->localsplus[0] = owner;
             for (int i = 1; i < code->co_nlocalsplus; i++) {
                 new_frame->localsplus[i] = NULL;