Specializes for
- Mutable class attributes with descriptors (e.g. enum.Enum members)
- `@property()` with Python functions