@Fidget-Spinner 2022-06-16
I'm quite surprised that the property specialisation's speedup is less than 2x:
(no PGO, no LTO, release build)
```
python -m timeit -s "
class X:
 y = property(lambda s: 10)

x = X()
" "x.y"

# MAIN
5000000 loops, best of 5: 40.2 nsec per loop
5000000 loops, best of 5: 39 nsec per loop
5000000 loops, best of 5: 38.5 nsec per loop

# this PR
10000000 loops, best of 5: 27.5 nsec per loop
10000000 loops, best of 5: 27.2 nsec per loop
10000000 loops, best of 5: 27 nsec per loop
```
Maybe you work in #93897 shall fix that.