@markshannon 2022-06-16
Given that https://github.com/python/cpython/issues/93910 is likely to be messing up our stats, maybe put this hold for a while?
Or, if you want to make a PR with just the property specialization, we could merge that.