@Fidget-Spinner 2022-07-08
Thanks to #94614 and #93910 we don't need this anymore.